backend/requirements.txt
appnope==0.1.0
click==6.7
decorator==4.2.1
Flask==0.12.2
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
MarkupSafe==1.0
mysql-connector-python-rf==2.2.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
Flask-Cors==3.0.3
-e ./
